Moved delete listener from ServletListener to DO
ServletContextListeners will only run in Servlet containers, non Servlet
applications will not run this automatically. In this case this specific
deletion listener is needed by a non servlet app so we're moving it to
the domain class as static code since this code will always be ran in
the FF initialization.

Actually this should be the standard solution when implementing deletion
listeners.