Fix list of active authorizations not answered
Only trigger authorizations redirector for child authorization if root
authorization is active.

Basicaly the redirect was being triggered by an active child
authorization although the root authorzation isn't active.
This made the users go to the authorization screen and unable to
answer the authorizations because of the inactive root authorization.

This fixes #UL-ISEG-1954 and related